#856 Split TestGitRepository into multiple more coherent classes. More tests for FetchRemotes().
